Rename ElasticsearchClient class
